requirements.txt
requests==2.31.0
psycopg2==2.9.9
python_dotenv==1.0.0
sqlalchemy==2.0.22
humanize==4.8.0
